LPS-134215 Renames Validation.es.js
